+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog for `cleff`
 
+## Unreleased
+
+### Changed
+
+- `Eff`'s instances are now handwritten instead of derived via `ReaderT`
+
+### Added
+
+- `Eff` has convenience lifted instances for more classes, so that some operations can be automatically lifted.
+  These classes are: `Bounded`, `Num`, `Fractional`, `Floating` and `IsString`
+- `Eff` now has a `MonadZip` instance for the `MonadComprehensions` extension
+
 ## 0.2.0.0 (2022-02-06)
 
 ### Changed